// ClientApp/src/app/core/nswag/plims-generated.ts
export const API_BASE_URL = new InjectionToken<string>('API_BASE_URL');
// ...
export class AdminService {
    private http: HttpClient;
    private baseUrl: string;
    protected jsonParseReviver: ((key: string, value: any) => any) | undefined = undefined;

    constructor(@Inject(HttpClient) http: HttpClient, @Optional() @Inject(API_BASE_URL) baseUrl?: string) {
        this.http = http;
        this.baseUrl = baseUrl !== undefined && baseUrl !== null ? baseUrl : "";
    }

    getTree(): Observable<AdminTreeItemInfo[] | null> {
        let url_ = this.baseUrl + "/api/Admin/tree";
        url_ = url_.replace(/[?&]$/, "");

        let options_ : any = {
            observe: "response",
            responseType: "blob",
            headers: new HttpHeaders({
                "Accept": "application/json"
            })
        };

        return this.http.request("get", url_, options_).pipe(_observableMergeMap((response_ : any) => {
            return this.processGetTree(response_);
        })).pipe(_observableCatch((response_: any) => {
            if (response_ instanceof HttpResponseBase) {
                try {
                    return this.processGetTree(<any>response_);
                } catch (e) {
                    return <Observable<AdminTreeItemInfo[] | null>><any>_observableThrow(e);
                }
            } else
                return <Observable<AdminTreeItemInfo[] | null>><any>_observableThrow(response_);
        }));
    }

    protected processGetTree(response: HttpResponseBase): Observable<AdminTreeItemInfo[] | null> {
        const status = response.status;
        const responseBlob =
            response instanceof HttpResponse ? response.body :
            (<any>response).error instanceof Blob ? (<any>response).error : undefined;

        let _headers: any = {}; if (response.headers) { for (let key of response.headers.keys()) { _headers[key] = response.headers.get(key); }}
        if (status === 200) {
            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
            let result200: any = null;
            result200 = _responseText === "" ? null : <AdminTreeItemInfo[]>JSON.parse(_responseText, this.jsonParseReviver);
            return _observableOf(result200);
            }));
        } else if (status !== 200 && status !== 204) {
            return blobToText(responseBlob).pipe(_observableMergeMap(_responseText => {
            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
            }));
        }
        return _observableOf<AdminTreeItemInfo[] | null>(<any>null);
    }
}
// ...
export interface AdminTreeItemInfo {
    name?: string | null;
    url?: string | null;
    type?: AdminTreeItemType;
}

export enum AdminTreeItemType {
    Users = 0,
    Places = 1,
}